# You-book Project/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "유북 API 명세서",
    "description": "Node.js Swaager swagger-jsdoc 방식 RestFul API 클라이언트 UI"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/semi-save": {
      "post": {
        "summary": "자서전 책 정보를 임시 저장하거나 기존의 임시저장 내역 업데이트",
        "tags": [
          "SemiSave"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": {
                    "type": "string",
                    "description": "책 ID"
                  },
                  "bookTitle": {
                    "type": "string",
                    "description": "책 제목 (기본값 \"제목없는 자서전\")"
                  },
                  "coverImg": {
                    "type": "string",
                    "description": "커버 이미지 경로"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "기존 레코드가 성공적으로 업데이트됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "새 레코드가 성공적으로 삽입됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류 또는 DB 오류"
          }
        }
      }
    },
    "/saved-articles": {
      "get": {
        "summary": "사용자의 모든 임시 저장된 자서전 책 정보 가져오기",
        "tags": [
          "SemiSave"
        ],
        "responses": {
          "200": {
            "description": "임시 저장된 자서전 목록을 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "articles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string",
                            "description": "자서전 제목"
                          },
                          "savedAt": {
                            "type": "string",
                            "description": "저장된 날짜와 시간"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류 또는 DB 오류"
          }
        }
      }
    },
...
    "/upload": {
      "post": {
        "summary": "이미지 파일을 업로드하고 경로를 DB에 저장",
        "tags": [
          "Image"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "업로드할 이미지 파일"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "이미지가 성공적으로 업로드되고 경로가 DB에 저장됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "성공 여부"
                    },
                    "path": {
                      "type": "string",
                      "description": "저장된 이미지의 경로"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "DB 저장 오류 또는 서버 오류"
          }
        }
      }
    },
    "/upload_base64": {
      "post": {
        "summary": "base64 인코딩된 이미지 데이터를 업로드하고 저장",
        "tags": [
          "Image"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "base64로 인코딩된 이미지 데이터"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "base64 이미지가 성공적으로 저장됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "성공 여부"
                    },
                    "path": {
                      "type": "string",
                      "description": "저장된 이미지의 경로"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "파일 저장 오류 또는 서버 오류"
          }
        }
      }
    },
...
  },
...
}